requirements.txt
pydantic
pydantic-settings
websocket-client
python-binance
pandas
pandas_ta
requests
google-generativeai